requirements.txt
aiomysql==0.0.21
aiosqlite==0.16.0
click==7.1.2
fastapi==0.62.0
h11==0.11.0
iso8601==0.1.13
pydantic==1.7.3
PyMySQL==0.9.3
PyPika==0.44.1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.4
starlette==0.13.6
text-unidecode==1.3
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
uvicorn==0.13.0
